+++ frappe/website/doctype/blog_post/test_blog_post.py
@@ -4,7 +4,7 @@ from __future__ import unicode_literals
 import frappe
 import unittest
 
-from frappe.tests.test_website import set_request
+from frappe.utils import set_request
 from frappe.website.render import render
 from frappe.utils import random_string
